A new CompositeLoggingProvider added for work item 4844. Some internal changes had to be made to the Logger and LoggingProviderBase to make this possible. The static Logger facade checked the configured providers for circular references during initialization to prevent stack overflows when providers logged to their fallback provider. The CompositeLoggingProvider allows referencing multiple providers and it was important to be able to also check these references during initialization. To achieve this, two internal virtual methods were added to the LoggingProviderBase class and the circular reference checking code in the Logger class was refactored and extracted to it's own class: CircularReferenceFinder.
A new CompositeException was added for the CompositeLoggingProvider (work item 4844).
The CompositeLoggingProvider can log to multiple providers and logging to other providers will continue, even when a providers failed. This means that this could lead to multiple failing providers. To be able to throw an exception that contains multiple inner exceptions, the new CompositeException was created. Its no coincidence that this exception looks a lot like the AggregateException of .NET 4.0. We can't however use AggregateException for two reasons:
    1. .NET 4.0 is currently in beta.
    2. CuttingEdge.Logging must stay compatible with .NET 2.0 for some time to come. 
By copying the AggregateException API we'll be able to let CompositeException inherit from AggregateException in a later release. This way a future version of CompositeLoggingProvider will be able to throw an AggregateException while not introducing a breaking change in the library :-).